Add update DetailView to use api/hooks folder
