+++ packages/invoices/src/InboundShipment/Dashboard/Widget.tsx
@@ -4,22 +4,21 @@ import {
   Grid,
   PlusCircleIcon,
   StatsPanel,
-  useAuthState,
   useQuery,
   useTranslation,
   Widget,
 } from '@openmsupply-client/common';
-import { getInboundShipmentCountQueryFn } from './api';
-import { useInboundShipmentApi } from '../api';
+import { useInboundApi } from '../api';
 
 export const InboundShipmentWidget: React.FC = () => {
-  const api = useInboundShipmentApi();
-  const { storeId } = useAuthState();
+  const api = useInboundApi();
   const t = useTranslation(['app', 'dashboard']);
   const { data, isLoading } = useQuery(
     ['inbound-shipment', 'count'],
-    getInboundShipmentCountQueryFn(api, storeId),
-    { retry: false }
+    api.dashboard.shipmentCount,
+    {
+      retry: false,
+    }
   );
 
   return (
