+++ packages/invoices/src/InboundShipment/DetailView/modals/InboundLineEdit/InboundLineEdit.tsx
@@ -26,7 +26,7 @@ import {
   useInboundFields,
   useSaveInboundLines,
   useNextItem,
-} from '../../api';
+} from '../../../api';
 import { InboundLineEditPanel } from './InboundLineEditPanel';
 
 interface InboundLineEditProps {
